fix: keep schema inspection out of lightweight version imports
Keep the version reader byte-identical to its prior lightweight contract. Header inspection retains Kysely and the deferred transaction owner, with read-only PRAGMAs owned by the existing SQLite configuration boundary. Full neutral Linux package build, native lock tests, and large-store proof pass without changing budgets or allowlists.
